Refactoring to use tmpl.go
