+++ anki/entry.go
@@ -3,42 +3,9 @@ package anki
 import (
 	"bytes"
 	"html/template"
-	"io/ioutil"
 	"log"
 	"strings"
-
-	"github.com/rakyll/statik/fs"
-	// https://github.com/rakyll/statik#usage
-	_ "github.com/takkyuuplayer/go-anki/anki/statik"
 )
-
-//go:generate statik -src assets -f
-
-var tmpl = template.New("entry")
-
-func init() {
-	mustParseAssets(tmpl, "/entry.html.tmpl")
-}
-
-func mustParseAssets(tmpl *template.Template, path string) *template.Template {
-	statikFS, err := fs.New()
-	if err != nil {
-		panic(err)
-	}
-
-	r, err := statikFS.Open(path)
-	if err != nil {
-		panic(err)
-	}
-	defer r.Close()
-
-	contents, err := ioutil.ReadAll(r)
-	if err != nil {
-		panic(err)
-	}
-
-	return template.Must(tmpl.Parse(string(contents)))
-}
 
 type Entry struct {
 	ID              string
@@ -74,7 +41,7 @@ type Accent struct {
 func (entry *Entry) AnkiCard() (string, error) {
 	buf := bytes.NewBufferString("")
 
-	if err := tmpl.Execute(buf, entry); err != nil {
+	if err := tmpl.Lookup("entry").Execute(buf, entry); err != nil {
 		log.Fatalf("execution failed: %s", err)
 		return "", err
 	}
